refactor(server): external-role-sync/sync.ts を external-role-sync.ts にフラット化
conditions.ts 削除後にディレクトリに 1 ファイルしか残らなくなったため。
